LayoutTests/http/tests/inspector/network/set-resource-caching-disabled-disk-cache-expected.txt

Test for `Network.setResourceCachingDisabled` for disk cache loads.


== Running test suite: Network.SetResourceCachingDisabled.DiskCache
-- Running test setup.
-- Running test case: PossibleNetworkLoad
PASS: Resource should be created.
PASS: Resource should receive a Response.
-- Running test setup.

-- Running test case: SetResourceCachingDisabled.DiskCache
PASS: Resource should be created.
PASS: Resource should receive a Response.
PASS: statusCode should be 200
PASS: responseSource should be Symbol(network)
